block, bfq-sq, bfq-mq: specify usage condition of delta_us in bfq_log_bfqq call
Inside the function bfq_completed_request the value of a variable
called delta_us is computed as current request completion time.
delta_us is used inside a call to the function bfq_log_bfqq as divisor
in a division operation to compute a rate value, but no check makes
sure that delta_us has non-zero value. A divisor with value 0 leads
to a division error that could result in a kernel oops (therefore
unstable/unreliable system state) and consequently cause kernel panic
if resources are unavailable after the system fault.

This commit fixes this call to bfq_log_bfqq specifying the condition
that allows delta_us to be safely used as divisor.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Chiara Bruschi <bruschi.chiara@outlook.it>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>